requirement.txt
llama-cpp-python
faster-whisper
sounddevice
scipy
numpy
pyttsx3
gradio